fix int -> long

+++ src/com/JJ/dao/jpa/RoleDAO.java
@@ -7,6 +7,6 @@ import org.springframework.transaction.annotation.Transactional;
 import com.JJ.TO.RoleTO;
 @Transactional(readOnly = true)
 public interface RoleDAO extends BaseDAO<RoleTO> {
-	RoleTO findByRoleId(Integer roleId);
-	List<RoleTO> findByRoleIdIn(List<Integer> roleIdList);
+	RoleTO findByRoleId(Long roleId);
+	List<RoleTO> findByRoleIdIn(List<Long> roleIdList);
 }
